Several recommended improvements
