+++ pom.xml
@@ -111,7 +111,5 @@
             <version>7.9.0</version>
             <scope>compile</scope>
         </dependency>
-
-
     </dependencies>
 </project>
