+++ src/main/java/org/com/epam/ta_training_java/AlionaSachok/driver/factory/Browser.java
@@ -1,0 +1,7 @@
+package org.com.epam.ta_training_java.AlionaSachok.driver.factory;
+
+import org.openqa.selenium.WebDriver;
+
+interface Browser {
+    WebDriver getBrowser();
+}
