+++ src/main/java/org/com/epam/ta_training_java/AlionaSachok/driver/factory/ChromeWithoutUI.java
@@ -1,0 +1,15 @@
+package org.com.epam.ta_training_java.AlionaSachok.driver.factory;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+public class ChromeWithoutUI implements Browser {
+    public WebDriver getBrowser() {
+        WebDriverManager.chromedriver().setup();
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("--headless");
+        return new ChromeDriver(options);
+    }
+}
