+++ src/main/java/org/com/epam/ta_training_java/AlionaSachok/driver/factory/FirefoxWithoutUI.java
@@ -1,0 +1,15 @@
+package org.com.epam.ta_training_java.AlionaSachok.driver.factory;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.firefox.FirefoxOptions;
+
+class FirefoxWithoutUI implements Browser {
+    public WebDriver getBrowser() {
+        WebDriverManager.firefoxdriver().setup();
+        FirefoxOptions options = new FirefoxOptions();
+        options.addArguments("--headless");
+        return new FirefoxDriver(options);
+    }
+}
