+++ src/test/resources/dev.properties
@@ -1,8 +1,8 @@
 base.url=https://www.saucedemo.com/
 
-browser.name=chrome_temporary
+#browser.name=chrome_temporary
 #user.name=standard_user
 #user.password=secret_sauce
-#browser.name=edge_temporary
+browser.name=edge_temporary
 #browser.name=firefox_temporary
-
+#browser.name=firefox_withoutui
