+++ src/test/resources/features/login.feature
@@ -10,9 +10,9 @@ Feature: Login Form Validation
     Then I should see an error message '<ErrorMessage>'
 
     Examples:
-      | Username         | Password     | ErrorMessage         |
-      | standard_user1   | secret_sauce | Username is required |
-      | locked_out_user1 | secret_sauce | Username is required |
+      | Username         | Password     | ErrorMessage                       |
+      | standard_user1   | secret_sauce | Epic sadface: Username is required |
+      | locked_out_user1 | secret_sauce | Epic sadface: Username is required |
 
   Scenario Outline: UC-2 Test Login Form with credentials by passing Username
     Given I am on the login page
@@ -23,9 +23,9 @@ Feature: Login Form Validation
     Then I should see an error message '<ErrorMessage>'
 
     Examples:
-      | Username    | Password     | ErrorMessage         |
-      | error_user  | secret_sauce | Password is required |
-      | visual_user | s            | Password is required |
+      | Username    | Password     | ErrorMessage                       |
+      | error_user  | secret_sauce | Epic sadface: Password is required |
+      | visual_user | s            | Epic sadface: Password is required |
 
   Scenario Outline: UC-3: Test Login Form with Credentials by Passing Username & Password
     Given I am on the login page
